SIRENA update (v.3.4.8)
- New SIRENA scripts in sixte/scripts/SIRENA:
  README, create_library_optfilt.sh, create_noise_spectrum.sh, example_reconstruction.sh, simulate_pulses_template.sh

-Changes to get the sampling rate (TESRECONSTRUCTION, GENNOISESPEC, testriggerfile.c):
   Sampling rate calculated by using DELTAT, TCLOCK+DEC_FAC or TCLOCK+NUMROW+P_ROW
   or read from the HISTORY keyword in the Primary unit
   or the XML file (the last one not in GENNOISESPEC).

   If not possible to get the sampling rate => Message asking to include DELTAT in the input FITS file

   In GENNOISESPEC, removed the input parameter 'samplingRate'

- GENNOISESPEC: - New 'rmNoiseIntervals' input parameter to clean (yes) or not the noise
                  (noise intervals whose sigma >meanSigmas+stdSigmas or <meanSigmas+stdSigmas are removed)
                - If scaleFactor!=0 => Low-pass filter applied not only to detection but also to noise spectrum
                - If I2R and new keywords ADU_CNV, ADU_BIAS and I_BIAS => R/R0 conversion

- TESRECONSTRUCTION: - intermediate=1 is already allowed to write pulses data in an intermediate FITS file
                       ('writePulses' in TASKSIRENA corrected)
                     - 'filterFile' input parameter deleted (not used, if intermediate=1 FILTER HDU is also written
                       in the intermediate FITS file)
                     - Able to read the grading from new input XML file:
                       <xifuconfig> <detector> </detector> <reconstruction> </reconstruction> </xifuconfig>

- INTEGRA: - Read ADU_CNV, ADU_BIAS and I_BIAS if I2R

- TASKSIRENA: - If I2R and new keywords ADU_CNV, ADU_BIAS and I_BIAS => R/R0 conversion
	      - New 'obtainRiseFallTimes' function (new columns RISETIME and FALLTIME in output FITS file)
	        Rise and fall times calculated with thresholds at 10% and 50%
              - 'writePulses' corrected

+++ tools/tesreconstruction/tesreconstruction.h
@@ -108,10 +108,10 @@ struct Parameters {
 	double scaleFactor;
 	
 	//Number of samples for threshold trespassing
-	double samplesUp;
+	int samplesUp;
         
         //Number of samples below the threshold to look for other pulse
-	double samplesDown;
+	int samplesDown;
 	
 	//Number of standard deviations in the kappa-clipping process for threshold estimation
 	double nSgms;
